replaced the bulky visual studio project with a lightweight g++ and makefile based project
