--- ppm-drawEngine/ppm-drawEngine.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30907.101
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ppm-drawEngine", "ppm-drawEngine\ppm-drawEngine.vcxproj", "{0708AFFC-08C3-4219-A7FE-30E74A2DF7B5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0708AFFC-08C3-4219-A7FE-30E74A2DF7B5}.Debug|x64.ActiveCfg = Debug|x64
-		{0708AFFC-08C3-4219-A7FE-30E74A2DF7B5}.Debug|x64.Build.0 = Debug|x64
-		{0708AFFC-08C3-4219-A7FE-30E74A2DF7B5}.Debug|x86.ActiveCfg = Debug|Win32
-		{0708AFFC-08C3-4219-A7FE-30E74A2DF7B5}.Debug|x86.Build.0 = Debug|Win32
-		{0708AFFC-08C3-4219-A7FE-30E74A2DF7B5}.Release|x64.ActiveCfg = Release|x64
-		{0708AFFC-08C3-4219-A7FE-30E74A2DF7B5}.Release|x64.Build.0 = Release|x64
-		{0708AFFC-08C3-4219-A7FE-30E74A2DF7B5}.Release|x86.ActiveCfg = Release|Win32
-		{0708AFFC-08C3-4219-A7FE-30E74A2DF7B5}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8AFF2B37-2197-4A44-BFAE-C8269D8938BF}
-	EndGlobalSection
-EndGlobal
